Move mock LLM service to test/llms; split into one class per file.
The mock LLM is a load-testing tool, not a production service component, so
relocate it under neuro_san/test/llms/. Split the single 422-line module
into one class per file to match the codebase convention.
